First pass of Damage Calculator
Damage Calculator that accounts for Affinity Bonus, Weakness Bonus, Panel Bonus, and rolls both Dice and Meteor.
Does not roll Dodge vs Accuracy yet.
Actual applications of the damage result not yet built.
This is the first part of a starter project I will be calling Battle Chip Showdown. It will be the core I build upon for the larger projects.